refactor: move models and create new domain
