+++ lib/core/migration/domain/interfaces/i_migration_repository.dart
@@ -1,4 +1,4 @@
-import 'package:tv_randshow/core/tvshow/domain/models/migration_status.dart';
+import 'package:tv_randshow/core/migration/domain/models/migration_status.dart';
 
 abstract class IMigrationRepository {
   Future<bool> saveStatus(MigrationStatus status);
