+++ lib/core/migration/domain/use_cases/save_migration_status_use_case.dart
@@ -1,7 +1,7 @@
 import 'package:injectable/injectable.dart';
 
-import 'package:tv_randshow/core/tvshow/domain/interfaces/i_migration_repository.dart';
-import 'package:tv_randshow/core/tvshow/domain/models/migration_status.dart';
+import 'package:tv_randshow/core/migration/domain/interfaces/i_migration_repository.dart';
+import 'package:tv_randshow/core/migration/domain/models/migration_status.dart';
 
 @injectable
 class SaveMigrationStatusUseCase {
