+++ lib/core/migration/domain/use_cases/verify_database_use_case.dart
@@ -1,7 +1,6 @@
 import 'package:injectable/injectable.dart';
-
-import 'package:tv_randshow/core/models/tvshow_details.dart';
 import 'package:tv_randshow/core/services/databases/i_database_service.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
 
 @Injectable()
 class VerifyDatabaseUseCase {
