+++ lib/core/services/databases/i_database_service.dart
@@ -1,4 +1,4 @@
-import 'package:tv_randshow/core/models/tvshow_details.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
 
 abstract class IDatabaseService {
   Future<void> saveTvshow(TvshowDetails tvshowDetails);
