+++ lib/core/services/databases/sql_database_service.dart
@@ -1,8 +1,8 @@
 import 'dart:developer';
 
 import 'package:injectable/injectable.dart';
-import 'package:tv_randshow/core/models/tvshow_details.dart';
 import 'package:tv_randshow/core/services/databases/i_secondary_database_service.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
 import 'package:tv_randshow/core/utils/database_helper.dart';
 
 @LazySingleton(as: ISecondaryDatabaseService, env: ['mobile'])
