+++ lib/core/services/favs_service.dart
@@ -3,12 +3,12 @@ import 'dart:ui' as ui;
 
 import 'package:injectable/injectable.dart';
 import 'package:tv_randshow/config/locator.dart';
-import 'package:tv_randshow/core/models/query.dart';
-import 'package:tv_randshow/core/models/tvshow_details.dart';
 import 'package:tv_randshow/core/services/databases/i_database_service.dart';
 import 'package:tv_randshow/core/streaming/domain/models/streaming_search.dart';
 import 'package:tv_randshow/core/streaming/domain/use_cases/get_tvshow_streamings_use_case.dart';
 import 'package:tv_randshow/core/tvshow/domain/interfaces/i_tvshow_repository.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/query.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
 
 @lazySingleton
 class FavsService {
