+++ lib/core/services/random_service.dart
@@ -2,13 +2,12 @@ import 'dart:developer' as developer;
 import 'dart:math';
 
 import 'package:injectable/injectable.dart';
-
-import 'package:tv_randshow/core/models/episode.dart';
-import 'package:tv_randshow/core/models/query.dart';
-import 'package:tv_randshow/core/models/tvshow_details.dart';
-import 'package:tv_randshow/core/models/tvshow_result.dart';
-import 'package:tv_randshow/core/models/tvshow_seasons_details.dart';
 import 'package:tv_randshow/core/tvshow/domain/interfaces/i_tvshow_repository.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/episode.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/query.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_result.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_seasons_details.dart';
 
 @lazySingleton
 class RandomService {
