+++ lib/core/tvshow/data/repositories/tvshow_repository.dart
@@ -1,10 +1,10 @@
 import 'package:injectable/injectable.dart';
-import 'package:tv_randshow/core/models/query.dart';
-import 'package:tv_randshow/core/models/search.dart';
-import 'package:tv_randshow/core/models/tvshow_details.dart';
-import 'package:tv_randshow/core/models/tvshow_seasons_details.dart';
 import 'package:tv_randshow/core/tvshow/data/services/tvshow_http_service.dart';
 import 'package:tv_randshow/core/tvshow/domain/interfaces/i_tvshow_repository.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/query.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/search.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_seasons_details.dart';
 
 @Injectable(as: ITvshowRepository)
 class TvshowRepository implements ITvshowRepository {
