+++ lib/core/tvshow/domain/interfaces/i_tvshow_repository.dart
@@ -1,7 +1,7 @@
-import 'package:tv_randshow/core/models/query.dart';
-import 'package:tv_randshow/core/models/search.dart';
-import 'package:tv_randshow/core/models/tvshow_details.dart';
-import 'package:tv_randshow/core/models/tvshow_seasons_details.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/query.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/search.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_seasons_details.dart';
 
 abstract class ITvshowRepository {
   Future<Search> search(Query query);
