+++ lib/core/tvshow/domain/models/query.dart
