+++ lib/core/tvshow/domain/models/tvshow_result.dart
@@ -1,4 +1,4 @@
-import 'package:tv_randshow/core/models/tvshow_details.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
 
 class TvshowResult {
   TvshowResult({
