+++ lib/core/tvshow/domain/use_cases/search_tvshow_use_case.dart
@@ -1,7 +1,7 @@
 import 'package:injectable/injectable.dart';
-import 'package:tv_randshow/core/models/query.dart';
-import 'package:tv_randshow/core/models/search.dart';
 import 'package:tv_randshow/core/tvshow/domain/interfaces/i_tvshow_repository.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/query.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/search.dart';
 
 @injectable
 class SearchTvShowUseCase {
