+++ lib/core/viewmodels/views/loading_view_model.dart
@@ -1,9 +1,8 @@
 import 'package:stacked/stacked.dart';
-
 import 'package:tv_randshow/config/locator.dart';
-import 'package:tv_randshow/core/models/tvshow_details.dart';
-import 'package:tv_randshow/core/models/tvshow_result.dart';
 import 'package:tv_randshow/core/services/random_service.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_result.dart';
 
 class LoadingViewModel extends BaseViewModel {
   final RandomService _randomService = locator<RandomService>();
