+++ lib/core/viewmodels/widgets/random_model.dart
@@ -1,8 +1,8 @@
 import 'package:stacked/stacked.dart';
 
 import 'package:tv_randshow/config/locator.dart';
-import 'package:tv_randshow/core/models/query.dart';
-import 'package:tv_randshow/core/models/tvshow_details.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/query.dart';
+import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
 import 'package:tv_randshow/core/tvshow/domain/use_cases/get_tvshow_details_use_case.dart';
 
 class RandomModel extends BaseViewModel {
